refactor(fft): share radix factorization and power-of-two check
Move the duplicated radix-8 preferred stage counter and the two
divergent isPowerOfTwo implementations into factorization.es.ts so
every caller runs the identical check. The c2r pipeline now rejects
non-factorable sizes explicitly instead of silently returning
truncated counts.
